+++ plugin/controllers/models/services.py
@@ -149,10 +149,13 @@ def getCurrentFullInfo(session):
 	frontendData = feinfo and feinfo.getAll(True)
 
 	if frontendData is not None:
+		cur_info = feinfo.getTransponderData(True)
 		inf['tunertype'] = frontendData.get("tuner_type", "UNKNOWN")
 		if frontendData.get("system", -1) == 1:
 			inf['tunertype'] = "DVB-S2"
 		inf['tunernumber'] = frontendData.get("tuner_number")
+		if cur_info:
+			inf['orbital_position'] = cur_info.get('orbital_position', None)
 	else:
 		inf['tunernumber'] = "N/A"
 		inf['tunertype'] = "N/A"
@@ -761,7 +764,7 @@ def getSearchSimilarEpg(ref, eventid):
 	return { "events": ret, "result": True }
 
 
-def getMultiEpg(self, ref, begintime=-1, endtime=None):
+def getMultiEpg(self, ref, begintime=-1, endtime=None, Mode=1):
 	# Check if an event has an associated timer. Unfortunately
 	# we cannot simply check against timer.eit, because a timer
 	# does not necessarily have one belonging to an epg event id.
@@ -828,18 +831,25 @@ def getMultiEpg(self, ref, begintime=-1, endtime=None):
 			ev['shortdesc'] = convertDesc(event[3])
 			ev['ref'] = event[4]
 			ev['timerStatus'] = getTimerEventStatus(event)
+			if Mode == 2:
+				ev['duration'] = event[6]
 
 			channel = filterName(event[5])
 			if not ret.has_key(channel):
-				ret[channel] = [ [], [], [], [], [], [], [], [], [], [], [], [] ]
+				if Mode == 1:
+					ret[channel] = [ [], [], [], [], [], [], [], [], [], [], [], [] ]
+				else:
+					ret[channel] = [[]]
 				picons[channel] = getPicon(event[4])
 
-			slot = int((event[1]-offset) / 7200)
-			if slot < 0:
-				slot = 0
-			if slot < 12 and event[1] < lastevent:
-				ret[channel][slot].append(ev)
-
+			if Mode == 1:
+				slot = int((event[1]-offset) / 7200)
+				if slot < 0:
+					slot = 0
+				if slot < 12 and event[1] < lastevent:
+					ret[channel][slot].append(ev)
+			else:
+				ret[channel][0].append(ev)
 	return { "events": ret, "result": True, "picons": picons }
 
 def getPicon(sname):