+++ plugin/local.py
@@ -404,6 +404,7 @@ tstrings = {'mo': _("Mo"),
 	'on_id': _("Onid"),
 	's_id': _("Sid"),
 	's_ref': _("Sref"),
+	's_orb': _("Orbital Position"),
 
 	'tstr_error_load_page': _("error! Loading Page"),
 	'tstr_timer_added': _("Timer Added"),
@@ -414,5 +415,7 @@ tstrings = {'mo': _("Mo"),
 	'tstr_theme': _("Theme"),
 	'tstr_show_picon_in_channel_list': _("Show Picon in channel list"),
 	'tstr_spinner': _("Spinner"),
+	'tstr_mepg_mode': _("Multi EPG Mode"),
+	'display_duration_s': _("Display duration (s)"),
 
 	}